+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Deployment modal - License renamed to End-User License Agreement
 - Module plugin - License renamed to End-User License Agreement
 - Paginations - Items per page is stored to and loaded from localStorage #1009
+- Main - Remove deprecated message and bootstrap-message
 
 ## [2.32.11] - 2022-12-19